Fix other and interpolation
