+++ src/mock/film.js
@@ -1,13 +1,11 @@
 import {nanoid} from 'nanoid';
-import {generateComments} from './comments';
 import {getRandomInteger} from './utils';
 import {generateRating} from './utils';
 import {generateInfo} from './utils';
 import {choise} from './utils';
 import {generateDescription} from './utils';
 import {generateTime} from './utils';
-import {generateRandomComments} from './utils';
-
+import {generateRandomComments} from './comments';
 
 const REGISSEURS = [
   `Anthony Mann`,
@@ -61,16 +59,16 @@ export const generateFilm = () => {
     description: generateDescription(),
     time: generateTime(),
     date: new Date(getRandomInteger(0, new Date())),
-    genre: choise(genres, getRandomInteger(1, genres.length - 1)),
+    genre: choise(GENRES, getRandomInteger(1, GENRES.length - 1)),
     comments: generateRandomComments(),
     isFavorite: Boolean(getRandomInteger(0, 1)),
     isViewed: Boolean(getRandomInteger(0, 1)),
     isWatchlist: Boolean(getRandomInteger(0, 1)),
     rating: generateRating(0, 9),
     age: getRandomInteger(0, 18),
-    regisseur: choise(regisseurs, 1),
-    screenwriters: choise(screenwriters, getRandomInteger(1, screenwriters.length - 1)),
-    actors: choise(actors, getRandomInteger(1, actors.length - 1)),
-    country: choise(countries, getRandomInteger(1, countries.length - 1)),
+    regisseur: choise(REGISSEURS, 1),
+    screenwriters: choise(SCREENWRITERS, getRandomInteger(1, SCREENWRITERS.length - 1)),
+    actors: choise(ACTORS, getRandomInteger(1, ACTORS.length - 1)),
+    country: choise(COUNTRIES, getRandomInteger(1, COUNTRIES.length - 1)),
   };
 };
